Fixed: types warnings
git-svn-id: http://dev.ryzom.com/svn/trunk@50 60ea55b9-4480-4529-b8ac-ed5291525b0f

+++ nel/src/misc/time_nl.cpp
@@ -46,7 +46,7 @@ namespace NLMISC
  */
 uint32 CTime::getSecondsSince1970 ()
 {
-	return nl_time (NULL);
+	return (uint32)nl_time (NULL);
 }
 
 /** Return the number of second since midnight (00:00:00), January 1, 1970,
